+++ src/ast/irbuilderbpf.cpp
@@ -73,14 +73,14 @@ libbpf::bpf_func_id IRBuilderBPF::selectProbeReadHelper(AddrSpace as, bool str)
 // It represents the inode of the initial (global) PID namespace
 constexpr uint32_t PROC_PID_INIT_INO = 0xeffffffc;
 
-Value *IRBuilderBPF::CreateGetPid(Value *ctx, const Location &loc)
+Value *IRBuilderBPF::CreateGetPid(const Location &loc)
 {
   const auto &pidns = bpftrace_.get_pidns_self_stat();
   if (pidns && pidns->st_ino != PROC_PID_INIT_INO) {
     // Get namespaced target PID when we're running in a namespace
     AllocaInst *res = CreateAllocaBPF(BpfPidnsInfoType(), "bpf_pidns_info");
     CreateGetNsPidTgid(
-        ctx, getInt64(pidns->st_dev), getInt64(pidns->st_ino), res, loc);
+        getInt64(pidns->st_dev), getInt64(pidns->st_ino), res, loc);
     Value *pid = CreateLoad(
         getInt32Ty(),
         CreateGEP(BpfPidnsInfoType(), res, { getInt32(0), getInt32(0) }));
@@ -94,14 +94,14 @@ Value *IRBuilderBPF::CreateGetPid(Value *ctx, const Location &loc)
   return pid;
 }
 
-Value *IRBuilderBPF::CreateGetTid(Value *ctx, const Location &loc)
+Value *IRBuilderBPF::CreateGetTid(const Location &loc)
 {
   const auto &pidns = bpftrace_.get_pidns_self_stat();
   if (pidns && pidns->st_ino != PROC_PID_INIT_INO) {
     // Get namespaced target TID when we're running in a namespace
     AllocaInst *res = CreateAllocaBPF(BpfPidnsInfoType(), "bpf_pidns_info");
     CreateGetNsPidTgid(
-        ctx, getInt64(pidns->st_dev), getInt64(pidns->st_ino), res, loc);
+        getInt64(pidns->st_dev), getInt64(pidns->st_ino), res, loc);
     Value *tid = CreateLoad(
         getInt32Ty(),
         CreateGEP(BpfPidnsInfoType(), res, { getInt32(0), getInt32(1) }));
@@ -115,8 +115,7 @@ Value *IRBuilderBPF::CreateGetTid(Value *ctx, const Location &loc)
   return tid;
 }
 
-AllocaInst *IRBuilderBPF::CreateUSym(Value *ctx,
-                                     Value *val,
+AllocaInst *IRBuilderBPF::CreateUSym(Value *val,
                                      int probe_id,
                                      const Location &loc)
 {
@@ -128,7 +127,7 @@ AllocaInst *IRBuilderBPF::CreateUSym(Value *ctx,
   StructType *usym_t = GetStructType("usym_t", elements, false);
   AllocaInst *buf = CreateAllocaBPF(usym_t, "usym");
 
-  Value *pid = CreateGetPid(ctx, loc);
+  Value *pid = CreateGetPid(loc);
   Value *probe_id_val = Constant::getIntegerValue(getInt32Ty(),
                                                   APInt(32, probe_id));
 
@@ -726,22 +725,18 @@ CallInst *IRBuilderBPF::createGetScratchMap(const std::string &map_name,
   return call;
 }
 
-Value *IRBuilderBPF::CreateMapLookupElem(Value *ctx,
-                                         Map &map,
+Value *IRBuilderBPF::CreateMapLookupElem(Map &map,
                                          Value *key,
                                          const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
-  return CreateMapLookupElem(ctx, map.ident, key, map.value_type, loc);
+  return CreateMapLookupElem(map.ident, key, map.value_type, loc);
 }
 
-Value *IRBuilderBPF::CreateMapLookupElem(Value *ctx,
-                                         const std::string &map_name,
+Value *IRBuilderBPF::CreateMapLookupElem(const std::string &map_name,
                                          Value *key,
                                          SizedType &type,
                                          const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   CallInst *call = createMapLookup(map_name, key);
 
   // Check if result == 0
@@ -776,7 +771,7 @@ Value *IRBuilderBPF::CreateMapLookupElem(Value *ctx,
     CreateMemsetBPF(value, getInt8(0), type.GetSize());
   else
     CreateStore(getInt64(0), value);
-  CreateHelperError(ctx, getInt32(0), libbpf::BPF_FUNC_map_lookup_elem, loc);
+  CreateHelperError(getInt32(0), libbpf::BPF_FUNC_map_lookup_elem, loc);
   CreateBr(lookup_merge_block);
 
   SetInsertPoint(lookup_merge_block);
@@ -790,8 +785,7 @@ Value *IRBuilderBPF::CreateMapLookupElem(Value *ctx,
   return ret;
 }
 
-Value *IRBuilderBPF::CreatePerCpuMapAggElems(Value *ctx,
-                                             Map &map,
+Value *IRBuilderBPF::CreatePerCpuMapAggElems(Map &map,
                                              Value *key,
                                              const SizedType &type,
                                              const Location &loc)
@@ -811,8 +805,6 @@ Value *IRBuilderBPF::CreatePerCpuMapAggElems(Value *ctx,
   //   i++;
   // }
   // return ret;
-
-  assert(ctx && ctx->getType() == getPtrTy());
 
   const std::string &map_name = map.ident;
 
@@ -899,8 +891,7 @@ Value *IRBuilderBPF::CreatePerCpuMapAggElems(Value *ctx,
 
   SetInsertPoint(error_success_block);
 
-  CreateHelperError(
-      ctx, getInt32(0), libbpf::BPF_FUNC_map_lookup_percpu_elem, loc);
+  CreateHelperError(getInt32(0), libbpf::BPF_FUNC_map_lookup_percpu_elem, loc);
   CreateBr(while_end);
 
   SetInsertPoint(error_failure_block);
@@ -1087,15 +1078,14 @@ void IRBuilderBPF::createPerCpuAvg(AllocaInst *total,
   CreateStore(CreateAdd(count_val, CreateLoad(getInt64Ty(), count)), count);
 }
 
-void IRBuilderBPF::CreateMapUpdateElem(Value *ctx,
-                                       const std::string &map_ident,
+void IRBuilderBPF::CreateMapUpdateElem(const std::string &map_ident,
                                        Value *key,
                                        Value *val,
                                        const Location &loc,
                                        int64_t flags)
 {
   Value *map_ptr = GetMapVar(map_ident);
-  assert(ctx && ctx->getType() == getPtrTy());
+
   assert(key->getType()->isPointerTy());
   assert(val->getType()->isPointerTy());
 
@@ -1116,15 +1106,13 @@ void IRBuilderBPF::CreateMapUpdateElem(Value *ctx,
                               update_func,
                               { map_ptr, key, val, flags_val },
                               "update_elem");
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_map_update_elem, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_map_update_elem, loc);
 }
 
-void IRBuilderBPF::CreateMapDeleteElem(Value *ctx,
-                                       Map &map,
+void IRBuilderBPF::CreateMapDeleteElem(Map &map,
                                        Value *key,
                                        const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   assert(key->getType()->isPointerTy());
   Value *map_ptr = GetMapVar(map.ident);
 
@@ -1139,11 +1127,10 @@ void IRBuilderBPF::CreateMapDeleteElem(Value *ctx,
       delete_func_ptr_type);
   CallInst *call = createCall(
       delete_func_type, delete_func, { map_ptr, key }, "delete_elem");
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_map_delete_elem, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_map_delete_elem, loc);
 }
 
-Value *IRBuilderBPF::CreateForEachMapElem(Value *ctx,
-                                          Map &map,
+Value *IRBuilderBPF::CreateForEachMapElem(Map &map,
                                           Value *callback,
                                           Value *callback_ctx,
                                           const Location &loc)
@@ -1176,7 +1163,7 @@ Value *IRBuilderBPF::CreateForEachMapElem(Value *ctx,
         callback_ctx ? CreateIntToPtr(callback_ctx, getPtrTy()) : GetNull(),
         /*flags=*/getInt64(0) },
       "for_each_map_elem");
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_for_each_map_elem, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_for_each_map_elem, loc);
   return call;
 }
 
@@ -1301,14 +1288,12 @@ void IRBuilderBPF::CreateUnSetRecursion(const Location &loc)
   SetInsertPoint(merge_block);
 }
 
-void IRBuilderBPF::CreateProbeRead(Value *ctx,
-                                   Value *dst,
+void IRBuilderBPF::CreateProbeRead(Value *dst,
                                    llvm::Value *size,
                                    Value *src,
                                    AddrSpace as,
                                    const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   assert(size && size->getType()->getIntegerBitWidth() <= 32);
   size = CreateIntCast(size, getInt32Ty(), false);
 
@@ -1328,27 +1313,24 @@ void IRBuilderBPF::CreateProbeRead(Value *ctx,
                               proberead_func,
                               { dst, size, src },
                               probeReadHelperName(read_fn));
-  CreateHelperErrorCond(ctx, call, read_fn, loc);
+  CreateHelperErrorCond(call, read_fn, loc);
 }
 
-CallInst *IRBuilderBPF::CreateProbeReadStr(Value *ctx,
-                                           Value *dst,
+CallInst *IRBuilderBPF::CreateProbeReadStr(Value *dst,
                                            size_t size,
                                            Value *src,
                                            AddrSpace as,
                                            const Location &loc)
 {
-  return CreateProbeReadStr(ctx, dst, getInt32(size), src, as, loc);
+  return CreateProbeReadStr(dst, getInt32(size), src, as, loc);
 }
 
-CallInst *IRBuilderBPF::CreateProbeReadStr(Value *ctx,
-                                           Value *dst,
+CallInst *IRBuilderBPF::CreateProbeReadStr(Value *dst,
                                            llvm::Value *size,
                                            Value *src,
                                            AddrSpace as,
                                            const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   assert(size && size->getType()->isIntegerTy());
   if ([[maybe_unused]] auto *dst_alloca = dyn_cast<AllocaInst>(dst)) {
     assert(dst_alloca->getAllocatedType()->isArrayTy() &&
@@ -1372,7 +1354,7 @@ CallInst *IRBuilderBPF::CreateProbeReadStr(Value *ctx,
                               probereadstr_callee,
                               { dst, size_i32, src },
                               probeReadHelperName(read_fn));
-  CreateHelperErrorCond(ctx, call, read_fn, loc);
+  CreateHelperErrorCond(call, read_fn, loc);
   return call;
 }
 
@@ -1382,7 +1364,6 @@ Value *IRBuilderBPF::CreateUSDTReadArgument(Value *ctx,
                                             AddrSpace as,
                                             const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   // Argument size must be 1, 2, 4, or 8. See
   // https://sourceware.org/systemtap/wiki/UserSpaceProbeImplementation
   int abs_size = std::abs(argument->size);
@@ -1457,7 +1438,7 @@ Value *IRBuilderBPF::CreateUSDTReadArgument(Value *ctx,
       if (index_offset) {
         ptr = CreateAdd(ptr, index_offset);
       }
-      CreateProbeRead(ctx, dst, getInt32(abs_size), ptr, as, loc);
+      CreateProbeRead(dst, getInt32(abs_size), ptr, as, loc);
       result = CreateLoad(getIntNTy(abs_size * 8), dst);
     } else {
       result = CreateLoad(getIntNTy(abs_size * 8), reg);
@@ -1478,7 +1459,6 @@ Value *IRBuilderBPF::CreateUSDTReadArgument(Value *ctx,
                                             AddrSpace as,
                                             const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   struct bcc_usdt_argument argument;
 
   void *usdt;
@@ -1781,8 +1761,7 @@ CallInst *IRBuilderBPF::CreateJiffies64(const Location &loc)
       libbpf::BPF_FUNC_jiffies64, jiffies64_func_type, {}, "jiffies64", loc);
 }
 
-Value *IRBuilderBPF::CreateIntegerArrayCmp(Value *ctx,
-                                           Value *val1,
+Value *IRBuilderBPF::CreateIntegerArrayCmp(Value *val1,
                                            Value *val2,
                                            const SizedType &val1_type,
                                            const SizedType &val2_type,
@@ -1854,8 +1833,7 @@ Value *IRBuilderBPF::CreateIntegerArrayCmp(Value *ctx,
   if (inBpfMemory(val1_type)) {
     val1_elem_i = CreateLoad(GetType(elem_type), ptr_val1_elem_i);
   } else {
-    CreateProbeRead(ctx,
-                    v1,
+    CreateProbeRead(v1,
                     getInt32(elem_type.GetSize()),
                     ptr_val1_elem_i,
                     val1_type.GetAS(),
@@ -1870,8 +1848,7 @@ Value *IRBuilderBPF::CreateIntegerArrayCmp(Value *ctx,
   if (inBpfMemory(val2_type)) {
     val2_elem_i = CreateLoad(GetType(elem_type), ptr_val2_elem_i);
   } else {
-    CreateProbeRead(ctx,
-                    v2,
+    CreateProbeRead(v2,
                     getInt32(elem_type.GetSize()),
                     ptr_val2_elem_i,
                     val2_type.GetAS(),
@@ -1911,8 +1888,7 @@ CallInst *IRBuilderBPF::CreateGetPidTgid(const Location &loc)
                           loc);
 }
 
-void IRBuilderBPF::CreateGetNsPidTgid(Value *ctx,
-                                      Value *dev,
+void IRBuilderBPF::CreateGetNsPidTgid(Value *dev,
                                       Value *ino,
                                       AllocaInst *ret,
                                       const Location &loc)
@@ -1936,8 +1912,7 @@ void IRBuilderBPF::CreateGetNsPidTgid(Value *ctx,
                                     { dev, ino, ret, getInt32(struct_size) },
                                     "get_ns_pid_tgid",
                                     loc);
-  CreateHelperErrorCond(
-      ctx, call, libbpf::BPF_FUNC_get_ns_current_pid_tgid, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_get_ns_current_pid_tgid, loc);
 }
 
 llvm::Type *IRBuilderBPF::BpfPidnsInfoType()
@@ -2028,8 +2003,6 @@ CallInst *IRBuilderBPF::CreateGetStack(Value *ctx,
                                        StackType stack_type,
                                        const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
-
   int flags = 0;
   if (ustack)
     flags |= (1 << 8);
@@ -2048,7 +2021,7 @@ CallInst *IRBuilderBPF::CreateGetStack(Value *ctx,
                                     { ctx, buf, stack_size, flags_val },
                                     "get_stack",
                                     loc);
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_get_stack, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_get_stack, loc);
   return call;
 }
 
@@ -2102,8 +2075,7 @@ CallInst *IRBuilderBPF::CreateThisCpuPtr(Value *var, const Location &loc)
                           loc);
 }
 
-void IRBuilderBPF::CreateGetCurrentComm(Value *ctx,
-                                        AllocaInst *buf,
+void IRBuilderBPF::CreateGetCurrentComm(AllocaInst *buf,
                                         size_t size,
                                         const Location &loc)
 {
@@ -2120,22 +2092,13 @@ void IRBuilderBPF::CreateGetCurrentComm(Value *ctx,
                                     { buf, getInt64(size) },
                                     "get_comm",
                                     loc);
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_get_current_comm, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_get_current_comm, loc);
 }
 
-void IRBuilderBPF::CreateOutput(Value *ctx,
-                                Value *data,
-                                size_t size,
-                                const Location &loc)
+void IRBuilderBPF::CreateOutput(Value *data, size_t size, const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   assert(data && data->getType()->isPointerTy());
-
-  if (bpftrace_.feature_->has_map_ringbuf()) {
-    CreateRingbufOutput(data, size, loc);
-  } else {
-    CreatePerfEventOutput(ctx, data, size, loc);
-  }
+  CreateRingbufOutput(data, size, loc);
 }
 
 void IRBuilderBPF::CreateRingbufOutput(Value *data,
@@ -2213,20 +2176,18 @@ void IRBuilderBPF::CreateAtomicIncCounter(const std::string &map_name,
   CreateLifetimeEnd(key);
 }
 
-void IRBuilderBPF::CreatePerCpuMapElemInit(Value *ctx,
-                                           Map &map,
+void IRBuilderBPF::CreatePerCpuMapElemInit(Map &map,
                                            Value *key,
                                            Value *val,
                                            const Location &loc)
 {
   AllocaInst *initValue = CreateAllocaBPF(val->getType(), "initial_value");
   CreateStore(val, initValue);
-  CreateMapUpdateElem(ctx, map.ident, key, initValue, loc, BPF_ANY);
+  CreateMapUpdateElem(map.ident, key, initValue, loc, BPF_ANY);
   CreateLifetimeEnd(initValue);
 }
 
-void IRBuilderBPF::CreatePerCpuMapElemAdd(Value *ctx,
-                                          Map &map,
+void IRBuilderBPF::CreatePerCpuMapElemAdd(Map &map,
                                           Value *key,
                                           Value *val,
                                           const Location &loc)
@@ -2261,37 +2222,11 @@ void IRBuilderBPF::CreatePerCpuMapElemAdd(Value *ctx,
 
   SetInsertPoint(lookup_failure_block);
 
-  CreatePerCpuMapElemInit(ctx, map, key, val, loc);
+  CreatePerCpuMapElemInit(map, key, val, loc);
 
   CreateBr(lookup_merge_block);
   SetInsertPoint(lookup_merge_block);
   CreateLifetimeEnd(value);
-}
-
-void IRBuilderBPF::CreatePerfEventOutput(Value *ctx,
-                                         Value *data,
-                                         size_t size,
-                                         const Location &loc)
-{
-  Value *map_ptr = GetMapVar(to_string(MapType::PerfEvent));
-
-  Value *flags_val = getInt64(BPF_F_CURRENT_CPU);
-  Value *size_val = getInt64(size);
-
-  // long bpf_perf_event_output(struct pt_regs *ctx, struct bpf_map *map,
-  //                            u64 flags, void *data, u64 size)
-  FunctionType *perfoutput_func_type = FunctionType::get(getInt64Ty(),
-                                                         { getPtrTy(),
-                                                           map_ptr->getType(),
-                                                           getInt64Ty(),
-                                                           data->getType(),
-                                                           getInt64Ty() },
-                                                         false);
-  CreateHelperCall(libbpf::BPF_FUNC_perf_event_output,
-                   perfoutput_func_type,
-                   { ctx, map_ptr, flags_val, data, size_val },
-                   "perf_event_output",
-                   loc);
 }
 
 void IRBuilderBPF::CreateDebugOutput(std::string fmt_str,
@@ -2332,7 +2267,7 @@ void IRBuilderBPF::CreateTracePrintk(Value *fmt_ptr,
                    loc);
 }
 
-void IRBuilderBPF::CreateSignal(Value *ctx, Value *sig, const Location &loc)
+void IRBuilderBPF::CreateSignal(Value *sig, const Location &loc)
 {
   // long bpf_send_signal(u32 sig)
   // Return: 0 or error
@@ -2345,7 +2280,7 @@ void IRBuilderBPF::CreateSignal(Value *ctx, Value *sig, const Location &loc)
       getInt64(libbpf::BPF_FUNC_send_signal),
       signal_func_ptr_type);
   CallInst *call = createCall(signal_func_type, signal_func, { sig }, "signal");
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_send_signal, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_send_signal, loc);
 }
 
 void IRBuilderBPF::CreateOverrideReturn(Value *ctx, Value *rc)
@@ -2519,12 +2454,10 @@ static bool return_zero_if_err(libbpf::bpf_func_id func_id)
   return false;
 }
 
-void IRBuilderBPF::CreateHelperError(Value *ctx,
-                                     Value *return_value,
+void IRBuilderBPF::CreateHelperError(Value *return_value,
                                      libbpf::bpf_func_id func_id,
                                      const Location &loc)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   assert(return_value && return_value->getType() == getInt32Ty());
 
   if (bpftrace_.helper_check_level_ == 0 ||
@@ -2552,19 +2485,17 @@ void IRBuilderBPF::CreateHelperError(Value *ctx,
 
   const auto &layout = module_.getDataLayout();
   auto struct_size = layout.getTypeAllocSize(helper_error_struct);
-  CreateOutput(ctx, buf, struct_size, loc);
+  CreateOutput(buf, struct_size, loc);
   CreateLifetimeEnd(buf);
 }
 
 // Report error if a return value < 0 (or return value == 0 if compare_zero is
 // true)
-void IRBuilderBPF::CreateHelperErrorCond(Value *ctx,
-                                         Value *return_value,
+void IRBuilderBPF::CreateHelperErrorCond(Value *return_value,
                                          libbpf::bpf_func_id func_id,
                                          const Location &loc,
                                          bool compare_zero)
 {
-  assert(ctx && ctx->getType() == getPtrTy());
   if (bpftrace_.helper_check_level_ == 0 ||
       (bpftrace_.helper_check_level_ == 1 && return_zero_if_err(func_id)))
     return;
@@ -2587,13 +2518,12 @@ void IRBuilderBPF::CreateHelperErrorCond(Value *ctx,
     condition = CreateICmpSGE(ret, Constant::getNullValue(ret->getType()));
   CreateCondBr(condition, helper_merge_block, helper_failure_block);
   SetInsertPoint(helper_failure_block);
-  CreateHelperError(ctx, ret, func_id, loc);
+  CreateHelperError(ret, func_id, loc);
   CreateBr(helper_merge_block);
   SetInsertPoint(helper_merge_block);
 }
 
-void IRBuilderBPF::CreatePath(Value *ctx,
-                              Value *buf,
+void IRBuilderBPF::CreatePath(Value *buf,
                               Value *path,
                               Value *sz,
                               const Location &loc)
@@ -2607,7 +2537,7 @@ void IRBuilderBPF::CreatePath(Value *ctx,
                                     { path, buf, sz },
                                     "d_path",
                                     loc);
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_d_path, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_d_path, loc);
 }
 
 void IRBuilderBPF::CreateSeqPrintf(Value *ctx,
@@ -2644,7 +2574,7 @@ void IRBuilderBPF::CreateSeqPrintf(Value *ctx,
                               seq_printf_func,
                               { seq, fmt, fmt_size, data, data_len },
                               "seq_printf");
-  CreateHelperErrorCond(ctx, call, libbpf::BPF_FUNC_seq_printf, loc);
+  CreateHelperErrorCond(call, libbpf::BPF_FUNC_seq_printf, loc);
 }
 
 StoreInst *IRBuilderBPF::createAlignedStore(Value *val,
@@ -2654,8 +2584,7 @@ StoreInst *IRBuilderBPF::createAlignedStore(Value *val,
   return CreateAlignedStore(val, ptr, MaybeAlign(align));
 }
 
-void IRBuilderBPF::CreateProbeRead(Value *ctx,
-                                   Value *dst,
+void IRBuilderBPF::CreateProbeRead(Value *dst,
                                    const SizedType &type,
                                    Value *src,
                                    const Location &loc,
@@ -2664,7 +2593,7 @@ void IRBuilderBPF::CreateProbeRead(Value *ctx,
   AddrSpace as = addrSpace ? addrSpace.value() : type.GetAS();
 
   if (!type.IsPtrTy()) {
-    CreateProbeRead(ctx, dst, getInt32(type.GetSize()), src, as, loc);
+    CreateProbeRead(dst, getInt32(type.GetSize()), src, as, loc);
     return;
   }
 
@@ -2684,7 +2613,7 @@ void IRBuilderBPF::CreateProbeRead(Value *ctx,
   if (ptr_size != type.GetSize())
     CreateMemsetBPF(dst, getInt8(0), type.GetSize());
 
-  CreateProbeRead(ctx, dst, getInt32(ptr_size), src, as, loc);
+  CreateProbeRead(dst, getInt32(ptr_size), src, as, loc);
 }
 
 llvm::Value *IRBuilderBPF::CreateDatastructElemLoad(